Bugfix 6210: The console entries were reversed for lost passwd reset/serial.
The last of the "console=..." entries in grub's "linux" config line is
the "default console", and that's the one that the lost passwd reset
script will talk to.  For serial console, that needs to be:
console=ttyS0,9600.
(cherry picked from commit d310b152955bd45808d7724d86fb404e7a8a82cf)