Move acpid, acpi-support-base and cpufrequtils to a new non xen list
Including the acpid, acpi-support-base and cpufrequtils packages in the
XenSource image causes a number of errors to be logged to the console
on boot up, due to either missing devices or access permissions on this
platform.  Removing these packages from the dependencies list in the
vyatta-cfg-system package and adding them to a chroot_local-packagelist
instead allows them to be excluded in these images, while still
maintaining them in non XenSource images.