build-iso: update 05-event_tty to only enable console when ttyS0 exists
Update the chroot_local-hooks 05-event_tty so that it only attempts to
enable console access on ttyS0 if the device is actually present.  This
is a failsafe in case console access is accidentally removed, possibly
rendering a system unaccessible.  The console is only enabled on ttyS0
if no other configuration options have been set for the console.

Bug #544 http://bugzilla.vyos.net/show_bug.cgi?id=544